move svg to assets directory
