+++ src/app/globals.css
@@ -63,4 +63,6 @@
     @apply -mx-2 mt-2 space-y-1;
 }
 
-
+.loading {
+    @apply mr-2 h-4 w-4 animate-spin;
+}
