+++ src/components/ui/button/Button.tsx
@@ -6,9 +6,10 @@ import {Utils} from "@/lib/utils";
 
 const Button:FC <ButtonProps> = (args: ButtonProps)=>{
     const {variant, size, className, isLoading, children, ...props} = args
-    const cName = Utils.buttonClassNames({variant, size, className})
+    const cName = Utils.buttonClassNames({variant, size, className});
+
     return <button className={cName} disabled={isLoading} {...props}>
-        {isLoading ? <Loader2 className='mr-2 h-4 w-4 animate-spin'/> :null}
+        {isLoading ? <Loader2 className='loading'/> :null}
         {children}
     </button>
 }
